chore(ml-dsa/hax): drop casts on `gamma1_exponent`

+++ libcrux-ml-dsa/src/simd/portable/encoding/gamma1.rs
@@ -56,7 +56,7 @@ fn serialize_when_gamma1_is_2_pow_19(simd_unit: &Coefficients, serialized: &mut 
 
 #[inline(always)]
 pub(crate) fn serialize(simd_unit: &Coefficients, serialized: &mut [u8], gamma1_exponent: usize) {
-    match gamma1_exponent as u8 {
+    match gamma1_exponent {
         17 => serialize_when_gamma1_is_2_pow_17(simd_unit, serialized),
         19 => serialize_when_gamma1_is_2_pow_19(simd_unit, serialized),
         _ => unreachable!(),
@@ -130,7 +130,7 @@ fn deserialize_when_gamma1_is_2_pow_19(serialized: &[u8], simd_unit: &mut Coeffi
 
 #[inline(always)]
 pub(crate) fn deserialize(serialized: &[u8], out: &mut Coefficients, gamma1_exponent: usize) {
-    match gamma1_exponent as u8 {
+    match gamma1_exponent {
         17 => deserialize_when_gamma1_is_2_pow_17(serialized, out),
         19 => deserialize_when_gamma1_is_2_pow_19(serialized, out),
         _ => unreachable!(),
